data is now floats

+++ src/compaction.h
@@ -5,6 +5,6 @@ static const int BLOCK_SIZE = 1024;
 extern float timing;
 
 
-int *prefix_sum_naive(const int len, const int *in);
+float *prefix_sum_naive(const int len, const float *in);
 
-int *prefix_sum(const int len, const int *in);
+float *prefix_sum(const int len, const float *in);
