+++ src/prefix_sum.cpp
@@ -1,9 +1,9 @@
 #include "prefix_sum.h"
 
 
-void prefix_sum_cpu(const int len, const int *in, int *out)
+void prefix_sum_cpu(const int len, const float *in, float *out)
 {
-    int sum = 0;
+    float sum = 0;
     for (int i = 0; i < len; ++i) {
         out[i] = sum;
         sum += in[i];
