+++ src/prefix_sum.h
@@ -1,4 +1,4 @@
 #pragma once
 
 
-void prefix_sum_cpu(const int len, const int *in, int *out);
+void prefix_sum_cpu(const int len, const float *in, float *out);
